Add equality, non-equality, and conversion tests
