+++ tests/compute/sympy/test_equal.py
@@ -1,0 +1,91 @@
+# Copyright (c) 2019-2024, Jonas Eschle, Jim Pivarski, Eduardo Rodrigues, and Henry Schreiner.
+#
+# Distributed under the 3-clause BSD license, see accompanying file LICENSE
+# or https://github.com/scikit-hep/vector for details.
+
+from __future__ import annotations
+
+import sympy
+
+import vector
+
+x, y, z, t = sympy.symbols("x y z t")
+
+
+def test_planar_sympy():
+    v1 = vector.VectorSympy2D(azimuthal=vector.backends.sympy.AzimuthalSympyXY(x, y))
+    v2 = vector.VectorSympy2D(azimuthal=vector.backends.sympy.AzimuthalSympyXY(x, y))
+    assert v1.equal(v2)
+
+    for t1 in "xy", "rhophi":
+        for t2 in "xy", "rhophi":
+            transformed1, transformed2 = (
+                getattr(v1, "to_" + t1)(),
+                getattr(v2, "to_" + t2)(),
+            )
+            transformed1.equal(transformed2)
+
+
+def test_spatial_sympy():
+    v1 = vector.VectorSympy3D(
+        azimuthal=vector.backends.sympy.AzimuthalSympyXY(x, y),
+        longitudinal=vector.backends.sympy.LongitudinalSympyZ(z),
+    )
+    v2 = vector.VectorSympy3D(
+        azimuthal=vector.backends.sympy.AzimuthalSympyXY(x, y),
+        longitudinal=vector.backends.sympy.LongitudinalSympyZ(z),
+    )
+    assert v1.equal(v2)
+
+    for t1 in "xyz", "xytheta", "xyeta", "rhophiz", "rhophitheta", "rhophieta":
+        for t2 in "xyz", "xytheta", "xyeta", "rhophiz", "rhophitheta", "rhophieta":
+            transformed1, transformed2 = (
+                getattr(v1, "to_" + t1)(),
+                getattr(v2, "to_" + t2)(),
+            )
+            transformed1.equal(transformed2)
+
+
+def test_lorentz_sympy():
+    v1 = vector.VectorSympy4D(
+        azimuthal=vector.backends.sympy.AzimuthalSympyXY(x, y),
+        longitudinal=vector.backends.sympy.LongitudinalSympyZ(z),
+        temporal=vector.backends.sympy.TemporalSympyT(t),
+    )
+    v2 = vector.VectorSympy4D(
+        azimuthal=vector.backends.sympy.AzimuthalSympyXY(x, y),
+        longitudinal=vector.backends.sympy.LongitudinalSympyZ(z),
+        temporal=vector.backends.sympy.TemporalSympyT(t),
+    )
+    assert v1.equal(v2)
+
+    for t1 in (
+        "xyzt",
+        "xythetat",
+        "xyetat",
+        "rhophizt",
+        "rhophithetat",
+        "rhophietat",
+        "xyztau",
+        "xythetatau",
+        "xyetatau",
+        "rhophiztau",
+        "rhophithetatau",
+        "rhophietatau",
+    ):
+        for t2 in (
+            "xyzt",
+            "xythetat",
+            "xyetat",
+            "rhophizt",
+            "rhophithetat",
+            "rhophietat",
+            "xyztau",
+            "xythetatau",
+            "xyetatau",
+            "rhophiztau",
+            "rhophithetatau",
+            "rhophietatau",
+        ):
+            tr1, tr2 = getattr(v1, "to_" + t1)(), getattr(v2, "to_" + t2)()
+            tr1.equal(tr2)
